Updated AlgoGen.m to work with the GUI. Added comments for future changes. Added fs_opt.m to git tracking. Updated ignore to ignore *.mat
AlgoGen used to open a new figure h, and then plot to it. However, the GUI has axes defined for this function. Figure/sub-plots have been commented out, however a future version should update the options.*Axes fields when the GUI is not active.
fs_opt attempts to minimize the features given a threshold on the cost. It does this using an approach similar to the AIC. See: A note on Genetic Algorithms for large-scale feature selection (1989).