evaluate_par.m is finished - and should work though is untested (performance metrics also en route)
evaluate_par now uses spmd to have each core run the same function on different data. The data that changes is the parents matrix. The training/test data is pre-allocated space in AlgoGen.m and placed on each core. This should save on overhead and improve performance of the overall algorithm.